merge to master

+++ src/FHBingen/Bundle/MHBBundle/Controller/VerwaltungsController.php
@@ -8,22 +8,26 @@
 
 namespace FHBingen\Bundle\MHBBundle\Controller;
 
-use Doctrine\Common\Collections\ArrayCollection;
 use FHBingen\Bundle\MHBBundle\Entity;
 use FHBingen\Bundle\MHBBundle\Form;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\HttpFoundation\Response;
 
-
+/**
+ * Class VerwaltungsController
+ *
+ * @package FHBingen\Bundle\MHBBundle\Controller
+ */
 class VerwaltungsController extends Controller
 {
     /**
+     * Unterscheidet zwischen SGL und Dozent und gibt diese jeweils sortiert in einem Array zurück
+     *
+     * @return array
+     *
      * @Route("/restricted/sgl/showUsers", name="benutzerVerwaltung")
      * @Template("FHBingenMHBBundle:Verwaltung:benutzerVerwaltung.html.twig")
-     *
-     * Unterscheidet zwischen SGL und Dozent und gibt diese jeweils sortiert in einem Array zurück
      */
     public function SglShowUsersAction()
     {
@@ -35,7 +39,7 @@ class VerwaltungsController extends Controller
 
         foreach ($entries as $e) {
             //wenn nicht "Alle" oder "N.N." wird zwischen SGL und Dozent unterschieden und sortiert
-            if ($e->getName()!='Dummy') {
+            if ($e->getName() != 'Dummy') {
                 if (in_array('ROLE_SGL', $e->getRoles())) {
                     $sgl[] = $e;
                 } else {
@@ -53,10 +57,12 @@ class VerwaltungsController extends Controller
     }
 
     /**
-     * @Route("/restricted/sgl/passwordReset", name="passwdReset")
+     * Setzt das Passwort bei Dozenten auf "Autor"
+     * Setzt das Passwort bei SQL auf "Editor"
      *
-     * setzt das Passwort bei Dozenten auf "Autor"
-     * setzt das Passwort bei SQL auf "Editor"
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     *
+     * @Route("/restricted/sgl/passwordReset", name="passwdReset")
      */
     public function resetAction()
     {
@@ -80,60 +86,35 @@ class VerwaltungsController extends Controller
     }
 
     /**
-     * @Route("/restricted/sgl/createUsers", name="benutzerErstellen")
-     * @Template("FHBingenMHBBundle:Verwaltung:benutzerErstellen.html.twig")
-     *
      * Legt einen neuen Dozenten mit den eingegebenen Daten an
+     *
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     *
+     * @Route("/restricted/sgl/createUsers", name="benutzerErstellen")
      */
-
     public function SglCreateUserAction()
     {
-        //TODO: userCreate und updateUsers zusammenführen?
-        $dozent = new Entity\Dozent();
-        $form = $this->createForm(new Form\DozentType(), $dozent);
-
-        $request = $this->get('request');
-        $form->handleRequest($request);
-
-        if ($request->getMethod() == 'POST') {
-            if ($form->isValid()) {
-                $dozent->setAnrede($form->get('anrede')->getData());
-                $dozent->setTitel($form->get('titel')->getData());
-                $dozent->setName($form->get('name')->getData());
-                $dozent->setNachname($form->get('nachname')->getData());
-                $dozent->setEmail($form->get('email')->getData());
-                $dozent->setRole($form->get('roles')->getData());
-                if (in_array('ROLE_SGL', $dozent->getRoles())) {
-                    $dozent->setPassword('Editor');
-                }
-                if (in_array('ROLE_DOZENT', $dozent->getRoles())) {
-                    $dozent->setPassword('Autor');
-                }
-                $em = $this->getDoctrine()->getManager();
-                $em->persist($dozent);
-                $em->flush();
-
-                $this->get('session')->getFlashBag()->add('info', 'Der neue Nutzer wurde erfolgreich angelegt.');
-
-                return $this->redirect($this->generateUrl('benutzerVerwaltung'));
-            }
-        }
-
-        return array('form' => $form->createView(), 'pageTitle' => 'Nutzerverwaltung');
+        return $this->redirect($this->generateUrl('benutzerBearbeiten', array('userid' => -1)));
     }
 
     /**
+     * Updatet einen bestimmten Dozent anhand der angegebenen Daten
+     *
+     * @param int $userid
+     *
+     * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
+     *
      * @Route("/restricted/sgl/updateUsers/{userid}", name="benutzerBearbeiten")
      * @Template("FHBingenMHBBundle:Verwaltung:benutzerErstellen.html.twig")
-     *
-     * updatet einen bestimmten Dozent anhand der angegebenen Daten
      */
-
     public function SglUpdateUserAction($userid)
     {
-        //TODO: userCreate und updateUsers zusammenführen?
         $em = $this->getDoctrine()->getManager();
-        $dozent = $em->getRepository('FHBingenMHBBundle:Dozent')->findOneBy(array('Dozenten_ID' => $userid));
+        if ($userid == -1) {
+            $dozent = new Entity\Dozent();
+        } else {
+            $dozent = $em->getRepository('FHBingenMHBBundle:Dozent')->findOneBy(array('Dozenten_ID' => $userid));
+        }
         $form = $this->createForm(new Form\DozentType(), $dozent);
 
         $request = $this->get('request');
@@ -156,14 +137,17 @@ class VerwaltungsController extends Controller
                 /*
                  * TODO:trigger?
                  * Doku: Wenn SGL auf Dozent abgestuft wird, besitzt er immernoch den Studiengang, den er vorher hatte,
-                 * aber da er die rolle ROLE_SGL nicht mehr ausübt, kann er nicht mehr auf die SGl-Funktionen zugreifen,
+                 * aber da er die rolle ROLE_SGL nicht mehr ausübt, kann er nicht mehr auf die SGL-Funktionen zugreifen,
                  * da sich diese hinter der /restricted/sgl firewall befinden
                  */
 
                 $em->persist($dozent);
                 $em->flush();
-
-                $this->get('session')->getFlashBag()->add('info', 'Der Nutzer wurde erfolgreich bearbeitet.');
+                if ($userid == -1 ) {
+                    $this->get('session')->getFlashBag()->add('info', 'Der neue Nutzer wurde erfolgreich angelegt.');
+                } else {
+                    $this->get('session')->getFlashBag()->add('info', 'Der Nutzer wurde erfolgreich bearbeitet.');
+                }
 
                 return $this->redirect($this->generateUrl('benutzerVerwaltung'));
             }
@@ -173,20 +157,21 @@ class VerwaltungsController extends Controller
     }
 
     /**
+     * Sucht alle angelegten Studiengänge und gibt diese sortiert aus
+     *
+     * @return array
+     *
      * @Route("/restricted/sgl/showAllCourses", name="studiengangVerwaltung")
      * @Template("FHBingenMHBBundle:Verwaltung:alleStudiengaenge.html.twig")
-     *
-     * Sucht alle angelegten Studiengänge und gibt diese sortiert aus
      */
-
     public function SglShowAllCoursesAction()
     {
         $em = $this->getDoctrine()->getManager();
         $mhb = $em->createQuery(
-            'SELECT s.Studiengang_ID,s.Titel AS Studiengang,s.Grad,d.Titel, d.Nachname
+            'SELECT s.Studiengang_ID, s.Titel AS studiengangTitel, s.Grad, s.Kuerzel, d.Titel AS dozentTitel, d.Name, d.Nachname
             FROM  FHBingenMHBBundle:Studiengang s
             JOIN  FHBingenMHBBundle:Dozent d WITH  s.sgl =  d.Dozenten_ID
-            ORDER BY Studiengang ASC'
+            ORDER BY studiengangTitel ASC'
         );
         $result = $mhb->getResult();
 
@@ -194,77 +179,35 @@ class VerwaltungsController extends Controller
     }
 
     /**
-     * @Route("/restricted/sgl/createCourse", name="studiengangErstellen")
-     * @Template("FHBingenMHBBundle:Verwaltung:studiengangAnzeigen.html.twig")
-     *
      * Legt einen Studiengang anhand der angegebenen Daten an. Dazu gehören auch Fachgebiete und Vertiefungsrichtungen
+     *
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     *
+     * @Route("/restricted/sgl/createCourse", name="studiengangErstellen")
      */
     public function SglCreateCourseAction()
     {
-        //TODO: createCourse + updateCourse zusammenlegen?
-        $em = $this->getDoctrine()->getManager();
-        $studiengang = new Entity\Studiengang();
-        $form = $this->createForm(new Form\StudiengangType(), $studiengang);
-
-        $request = $this->get('request');
-        $form->handleRequest($request);
-
-        if ($request->getMethod() == 'POST') {
-            if ($form->isValid()) {
-
-                $studiengang->setFachbereich($form->get('fachbereich')->getData());     //choice
-                $studiengang->setGrad($form->get('grad')->getData());                   //choice
-                $studiengang->setTitel($form->get('titel')->getData());                 //text
-                $studiengang->setKuerzel($form->get('kuerzel')->getData());             //text
-                $studiengang->setBeschreibung($form->get('beschreibung')->getData());   //text
-                $studiengang->setSgl($form->get('sgl')->getData());                     //entity
-
-                /*
-                 * TODO:
-                 * - überprüfen ob Vertiefungsrichtung oder Fachgebiet doppelt in Feldern steht
-                 * - vllt sollte man Vertiefungen + Fachgebeiete nicht umbenennen können (oder nur über spezielle Maske)
-                 */
-
-                //hier keine Collection, sondern "nur" array
-                $vertiefungArr = $form->get('richtung')->getData();
-                foreach ($vertiefungArr as $vertiefung) {
-                    $studiengang->addRichtung($vertiefung);
-                    $vertiefung->setStudiengang($studiengang);
-                    $em->persist($vertiefung);
-                }
-
-                //hier keine Collection, sondern "nur" array
-                $fachgebietArr = $form->get('fachgebiete')->getData();
-                foreach ($fachgebietArr as $fachgebiet) {
-                    $studiengang->addFachgebiete($fachgebiet);
-                    $fachgebiet->setStudiengang($studiengang);
-                    $em->persist($fachgebiet);
-                }
-
-                $em->persist($studiengang);
-                $em->flush();
-
-                $this->get('session')->getFlashBag()->add('info', 'Der Studiengang wurde erfolgreich angelegt.');
-
-                return $this->redirect($this->generateUrl('studiengangVerwaltung'));
-            }
-        }
-
-        return array('form' => $form->createView(),'studiengang'=>$studiengang, 'pageTitle' => 'Studiengangverwaltung');
+        return $this->redirect($this->generateUrl('studiengangBearbeiten', array('courseID' => -1)));
     }
 
     /**
+     * Updatet einen Studiengang nach den angegebenen Daten
+     *
+     * @param int $courseID
+     *
+     * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
+     *
      * @Route("/restricted/sgl/updateCourse/{courseID}", name="studiengangBearbeiten")
      * @Template("FHBingenMHBBundle:Verwaltung:studiengangAnzeigen.html.twig")
-     *
-     * Updatet einen Studiengang nach den angegebenen Daten
      */
-    public function SglShowCourseAction($courseID)
+    public function SglUpdateCourseAction($courseID)
     {
-        //TODO: createCourse + updateCourse zusammenlegen?
-
         $em = $this->getDoctrine()->getManager();
-        $studiengang = $em->getRepository('FHBingenMHBBundle:Studiengang')->findOneBy(array('Studiengang_ID' => $courseID));
+        if ($courseID == -1) {
+            $studiengang = new Entity\Studiengang();
+        } else {
+            $studiengang = $em->getRepository('FHBingenMHBBundle:Studiengang')->findOneBy(array('Studiengang_ID' => $courseID));
+        }
         $form = $this->createForm(new Form\StudiengangType(), $studiengang);
 
         $request = $this->get('request');
@@ -272,11 +215,6 @@ class VerwaltungsController extends Controller
 
         if ($request->getMethod() == 'POST') {
             if ($form->isValid()) {
-                /*
-                 * TODO:
-                 * - überprüfen ob Vertiefungsrichtung oder Fachgebeiet doppelt in Feldern steht
-                 * - vllt sollte man Vertiefungen + Fachgebeiete nicht umbenennen können (oder nur über spezielle Maske)
-                 */
                 $studiengang->setFachbereich($form->get('fachbereich')->getData());     //choice
                 $studiengang->setGrad($form->get('grad')->getData());                   //choice
                 $studiengang->setTitel($form->get('titel')->getData());                 //text
@@ -284,14 +222,21 @@ class VerwaltungsController extends Controller
                 $studiengang->setBeschreibung($form->get('beschreibung')->getData());   //text
                 $studiengang->setSgl($form->get('sgl')->getData());                     //entity
 
-                $vertiefungArr = $form->get('richtung')->getData()->toArray();
+                //benötigt, sonst Fehler
+                if ($courseID == -1) {
+                    $vertiefungArr = $form->get('richtung')->getData();
+                    $fachgebietArr = $form->get('fachgebiete')->getData();
+                } else {
+                    $vertiefungArr = $form->get('richtung')->getData()->toArray();
+                    $fachgebietArr = $form->get('fachgebiete')->getData()->toArray();
+                }
+
                 foreach ($vertiefungArr as $vertiefung) {
                     $studiengang->addRichtung($vertiefung);
                     $vertiefung->setStudiengang($studiengang);
                     $em->persist($vertiefung);
                 }
 
-                $fachgebietArr = $form->get('fachgebiete')->getData()->toArray();
                 foreach ($fachgebietArr as $fachgebiet) {
                     $studiengang->addFachgebiete($fachgebiet);
                     $fachgebiet->setStudiengang($studiengang);
@@ -302,31 +247,38 @@ class VerwaltungsController extends Controller
                 // hier notwendig da weiter unten alle Eintrage mit den aktuell gesetzten verglichen werden
                 $em->flush();
 
-                //$studiengang->getRichtung() holt sich die infos NICHT aus der db....
-                //also:
-                $vertiefungRepository = $em->getRepository('FHBingenMHBBundle:Vertiefung');
-                $dbVertiefungArr = $vertiefungRepository->findby(array('studiengang' => $courseID));
+                //hier ungleich -1
+                if ($courseID != -1) {
+                    //$studiengang->getRichtung() holt sich die infos NICHT aus der db....
+                    //also:
+                    $vertiefungRepository = $em->getRepository('FHBingenMHBBundle:Vertiefung');
+                    $dbVertiefungArr = $vertiefungRepository->findby(array('studiengang' => $courseID));
 
-                $fachgebietRepository = $em->getRepository('FHBingenMHBBundle:Fachgebiet');
-                $dbFachgebietArr = $fachgebietRepository->findby(array('studiengang' => $courseID));
+                    $fachgebietRepository = $em->getRepository('FHBingenMHBBundle:Fachgebiet');
+                    $dbFachgebietArr = $fachgebietRepository->findby(array('studiengang' => $courseID));
 
 
-                //löscht alle Vertiefungsrichtungen aus der DB, die nicht angegeben/gelöscht wurden
-                foreach ($dbVertiefungArr as $dbEntry) {
-                    if (!in_array($dbEntry, $vertiefungArr)) {
-                        $em->remove($dbEntry);
+                    //löscht alle Vertiefungsrichtungen aus der DB, die nicht angegeben/gelöscht wurden
+                    foreach ($dbVertiefungArr as $dbEntry) {
+                        if (!in_array($dbEntry, $vertiefungArr)) {
+                            $em->remove($dbEntry);
+                        }
                     }
+
+                    //löscht alle Fachgebiete aus der DB, die nicht angegeben/gelöscht wurden
+                    foreach ($dbFachgebietArr as $dbEntry) {
+                        if (!in_array($dbEntry, $fachgebietArr)) {
+                            $em->remove($dbEntry);
+                        }
+                    }
+                    $em->flush();
                 }
 
-                //löscht alle Fachgebiete aus der DB, die nicht angegeben/gelöscht wurden
-                foreach ($dbFachgebietArr as $dbEntry) {
-                    if (!in_array($dbEntry, $fachgebietArr)) {
-                        $em->remove($dbEntry);
-                    }
+                if ($courseID == -1) {
+                    $this->get('session')->getFlashBag()->add('info', 'Der Studiengang wurde erfolgreich angelegt.');
+                } else {
+                    $this->get('session')->getFlashBag()->add('info', 'Der Studiengang wurde erfolgreich bearbeitet.');
                 }
-                $em->flush();
-
-                $this->get('session')->getFlashBag()->add('info', 'Der Studiengang wurde erfolgreich bearbeitet.');
 
                 return $this->redirect($this->generateUrl('studiengangVerwaltung'));
             }
@@ -337,9 +289,13 @@ class VerwaltungsController extends Controller
 
 
     /**
-     * @Route("/restricted/sgl/SglUserDeactivation/{userid}", name="SglUserDeactivation")
-     *
      * Deaktiviert den User
+     *
+     * @param int $userid
+     *
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     *
+     * @Route("/restricted/sgl/SglUserDeactivation/{userid}", name="SglUserDeactivation")
      */
     public function SglUserDeactivationAction($userid)
     {
@@ -354,9 +310,13 @@ class VerwaltungsController extends Controller
     }
 
     /**
-     * @Route("/restricted/sgl/SglUserActivation/{userid}", name="SglUserActivation")
-     *
      * Aktiviert den User
+     *
+     * @param int $userid
+     *
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     *
+     * @Route("/restricted/sgl/SglUserActivation/{userid}", name="SglUserActivation")
      */
     public function SglUserActivationAction($userid)
     {
@@ -369,7 +329,4 @@ class VerwaltungsController extends Controller
 
         return $this->redirect($this->generateUrl('benutzerVerwaltung'));
     }
-
-
-
 }
